Player lb

+++ src/main/java/com/parkourcraft/parkour/data/stats/PlayerStats.java
@@ -30,7 +30,7 @@ public class PlayerStats {
     private int infinitePKScore = 0;
     private boolean inInfinitePK = false;
     private boolean eventParticipant = false;
-    private long totalLevelCompletions = 0;
+    private int totalLevelCompletions = 0;
     private Map<String, List<LevelCompletion>> levelCompletionsMap = new HashMap<>();
     private Map<String, Long> perks = new HashMap<>();
 
@@ -118,9 +118,9 @@ public class PlayerStats {
         return levelStartTime;
     }
 
-    public long getTotalLevelCompletions() { return totalLevelCompletions; }
+    public int getTotalLevelCompletions() { return totalLevelCompletions; }
 
-    public void setTotalLevelCompletions(long totalLevelCompletions) { this.totalLevelCompletions = totalLevelCompletions; }
+    public void setTotalLevelCompletions(int totalLevelCompletions) { this.totalLevelCompletions = totalLevelCompletions; }
 
     //
     // Spectator Section
